Rewrite custom monad in terms of transformers
The custom monad can be represented through a monad transformer and the derivations of its classes (save MonadFail) can be automated. Traditionally, ExceptT is on the outermost layer in order to allow for the state to be accessed regardless of failure or success (useful for debugging). The rewrite preserves semantics though, so it keeps in line with the design decision of the original author, which may require reconsidering, if debugging upon failure by examining state is required.